Feedback and fixes
