+++ apps-rendering/src/components/Standfirst/index.tsx
@@ -4,10 +4,10 @@ import { ArticleDesign, ArticleDisplay, ArticleSpecial } from '@guardian/libs';
 import type { Item } from 'item';
 import { getFormat } from 'item';
 import DeadBlogStandfirst from './DeadBlogStandfirst';
+import GalleryStandfirst from './GalleryStandfirst';
 import ImmersiveLabsStandfirst from './ImmersiveLabsStandfirst';
 import ImmersiveStandfirst from './ImmersiveStandfirst';
 import InterviewStandfirst from './InterviewStandfirst';
-import GalleryStandfirst from './GalleryStandfirst';
 import LabsStandfirst from './LabsStandfirst';
 import LiveBlogStandfirst from './LiveBlogStandfirst';
 import MediaStandfirst from './MediaStandfirst';
